// typechain-types/contracts/FundMe.ts
export interface FundMeInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "MINIMUM_USD"
      | "cheaperWithdraw"
      | "fund"
      | "i_owner"
      | "s_addressToAmountFunded"
      | "s_funders"
      | "s_priceFeed"
      | "withdraw"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "MINIMUM_USD",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "cheaperWithdraw",
    values?: undefined
  ): string;
  encodeFunctionData(functionFragment: "fund", values?: undefined): string;
  encodeFunctionData(functionFragment: "i_owner", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "s_addressToAmountFunded",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "s_funders",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "s_priceFeed",
    values?: undefined
  ): string;
  encodeFunctionData(functionFragment: "withdraw", values?: undefined): string;

  decodeFunctionResult(
    functionFragment: "MINIMUM_USD",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "cheaperWithdraw",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "fund", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "i_owner", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "s_addressToAmountFunded",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "s_funders", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "s_priceFeed",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "withdraw", data: BytesLike): Result;
}

export interface FundMe extends BaseContract {
  connect(runner?: ContractRunner | null): FundMe;
  waitForDeployment(): Promise<this>;

  interface: FundMeInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  MINIMUM_USD: TypedContractMethod<[], [bigint], "view">;

  cheaperWithdraw: TypedContractMethod<[], [void], "payable">;

  fund: TypedContractMethod<[], [void], "payable">;

  i_owner: TypedContractMethod<[], [string], "view">;

  s_addressToAmountFunded: TypedContractMethod<
    [arg0: AddressLike],
    [bigint],
    "view"
  >;

  s_funders: TypedContractMethod<[arg0: BigNumberish], [string], "view">;

  s_priceFeed: TypedContractMethod<[], [string], "view">;

  withdraw: TypedContractMethod<[], [void], "nonpayable">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "MINIMUM_USD"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "cheaperWithdraw"
  ): TypedContractMethod<[], [void], "payable">;
  getFunction(
    nameOrSignature: "fund"
  ): TypedContractMethod<[], [void], "payable">;
  getFunction(
    nameOrSignature: "i_owner"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "s_addressToAmountFunded"
  ): TypedContractMethod<[arg0: AddressLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "s_funders"
  ): TypedContractMethod<[arg0: BigNumberish], [string], "view">;
  getFunction(
    nameOrSignature: "s_priceFeed"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "withdraw"
  ): TypedContractMethod<[], [void], "nonpayable">;

  filters: {};
}